declarative-camera: Fix flash-controls not being synced with Camera
The declarative-camera has a bug where the flash/torch-controls do not
get hidden/shown correctly whenever the camera-device is changed, this
can be easily reproduced by selecting a device that supports flash, and
then switching to one that does not support flash. Additionally, the
logic for applying/reading the QCamera::flashMode over time is fragile.

This patch introduces logic to refresh the 'mIsFlashSupported' members
whenever the camera-device is changed. It also updates the logic
of the Switch components to use reactive bindings properly.

As a drive-by, we make the 'Camera' property a required one, so we can
avoid unnecessary null-checks.

Pick-to: 6.10
Change-Id: I643c2a55e91520fdb9d55c423e4865504e8224ee
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>