Apple: Pass CVPixelBufferRef as opaque pointer into av_buffer_create
Currently, when freeing the AVBuffer, we pass the CVPixelBufferRef
through the data-pointer parameter. This parameter is usually used for
pixel-data that is potentially read and written by FFmpeg.

This patch moves the logic to use the opaque-data pointer, making sure
it's not touched before we can free it.

Change-Id: Iafa16da8141eabfcce9e5bdb44154db35f7fe0e8
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit ee8aeae3a147993ad962d3e3727932947f3c8dcb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>