Remove some Obj-C forward declarations in QtMultimedia darwin classes
With C++ based forward declarations for Obj-C classes, we can run into
scenarios where we accidentally use these class in .cpp files. This
can be error-prone if we turn on ARC.

These files are no longer used in any .cpp files. This patch removes
the forward declarations and AVFoundation directly.

Pick-to: 6.10
Change-Id: Id1be3f75b4a377d5ee5f992cebbafe74d84aefc9
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>