Tests: window capture backend
tst_QWindowCaptureBackend uses a custom main function in order to
provide an overload to spawn a subprocess to show a window to capture.

This custom main function overload is rather fragile, we can avoid it by
using QTEST_MAIN, but re-defining the `main` function that it defines to
`testlib_main` and call that from our custom main function.

This simplifies the implementation quite a bit.

Change-Id: Ibf5447ae6798ea0f28956768b92219e2b75b7fbe
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>