tst_qaudiosink: Add test for different sample formats
We were missing a test for starting a QAudioSink for any sample formats
supported by a specific QAudioDevice. This would've caught a regression
in sample format conversion.

This patch adds a test to tst_qaudiosink that tries to start a sink for
each sample format supported by the default audio output device.

Pick-to: 6.10
Task-number: QTBUG-140472
Change-Id: Iaf8608e200e50a0209d5480d11794a3f3dc5b365
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>