Pipewire: improve support for virtual nodes
Some Pipewire plugins may expose virtual audio sources / sinks. We
previously did not enumerate them, as they did tnot contain a
"device.id" property. Technically we only parse the devices in order to
obtain the sysfs path.
So we make deviceSerial optional in the device monitor and the sysfs
path optional in the audio device.

Fixes: QTBUG-141250
Pick-to: 6.10
Change-Id: I15e388d3427b08e47d134caff78a147b7d503399
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>