Pipewire: QAudioContextManager - remove wrong assertion
The metadata listener had a wrong assertion: if the `key` is nullptr,
all metadata for the subjects are supposted to be cleared. We actually
make use of this inside handleDefaultMetadataObjectEvent. So the wrong
assertion should be removed

Change-Id: I188caa8e29991f6ae5a15f2b73932229d0dba489
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
(cherry picked from commit eac0af10268db8a37c719e68b54191a3aa8513cc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>