Camera, docs: Copy camera-device description to C++ documentation
When this documentation was updated, it was not also applied to the C++
documentation. This was likely an oversight.

This patch copies the description from QML to the C++ documentation.

As a drive-by, we fix links (Camera vs QCamera) to match the docs.

Change-Id: If559c4f9c515d582d5a2e195b3a06101ce52a028
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit 11f0a8846ea408a19af6b5cdf098faa86a46364d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>