Camera, docs: Document QML supportedFeatures as enum type
The QML documentation for Camera.supportedFeatures denotes its type as
'Features'. This type is only exposed in C++, and is treated as a
regular int on the QML side. This patch documents it as an enumerator.

Change-Id: I3260f48d98708db9c3427afdbc6a2570153b45bc
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Bror Wetlesen Vedeld <bror.vedeld@qt.io>
(cherry picked from commit 7e8d72d2feae9eb661f9acaa8dbf8a3e0444e547)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>