Camera, docs: Remove examples using customFocusPoint
This functionality is not supported, and cannot be supported until the
public API receives additions to support it. At present, there is no
FocusMode::CustomFocusPoint, and so there is no way to communicate
to the backend that we should be using the custom focus point. We
should not present it to application developers as a functionality
they should expect to work.

This patch removes the examples that reference this functionality.

Pick-to: 6.10
Change-Id: Ie9581978e98a91fd4676592c6f9bc84cf1b65bb3
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>