QAVFCameraBase: Add QCameraDevice parameter to updateSupportedFeatures
The function 'updateSupprtedFeatures' in the class QAVFCameraBase is
responsible for updating the property 'supportedFeatures' property in
QCamera. Internally it relies on the QPlatformCamera member
'm_cameraDevice' in order to read info from the physical device.

This is error-prone because we have to be careful that we always update
'm_cameraDevice' ahead of time. This patch makes us pass the device
we want to use explicitly.

Pick-to: 6.10
Task-number: QTBUG-141376
Change-Id: Ia1f76fc49424a6dda39b1781eec2af32cc44e6e3
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>