QAVFCameraBase: Add QCameraDevice param to updateCameraConfiguration
The function 'updateCameraConfiguration' in the class QAVFCameraBase is
responsible for updating various capabilities and applying properties
to the physical camera device. Internally it relies on the
QPlatformCamera member 'm_cameraDevice' in order to access the
physical device.

This is error-prone because we have to be careful that we always update
'm_cameraDevice' ahead of time. This patch makes us pass the device
we want to use explicitly.

Task-number: QTBUG-141376
Change-Id: Id063651b2bec270e4cc67f43464e0be7d710d550
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit df62da9e080c14fe88639e00088b6162021e10c8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>