Add private API for setting preferred static QRhi instance backend
For testing purposes, it would be useful to specify a specific backend
for thread local QRhi instances, even if another backend would be
available, such as OpenGL instead of Metal.

This patch adds private API as an environment variable to specify a QRhi
backend to create when required as a static QRhi instance.

Pick-to: 6.10 6.8
Change-Id: I6fb3aad4c76ce20b2c07348799a78c1129b1d807
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>