Camera, docs: Remove section about cancelling still photo captures
The documentation describes that application developers may cancel
on-going still photo captures with QImageCapture by requesting a new
capture. Since the introduction of the FFmpeg backend, this is no
longer true, since QFFmpegImageCapture is currently hardcoded to
support only a single capture being queued at a time and subsequent
requests are ignored until the capture is finished or cancelled by the
backend.

This patch removes this section from the documentation.

Pick-to: 6.10
Change-Id: Id58cf2d17f7fdae9cd4d6c847100fdada460855e
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>