Apple, camera: Fix typo in camera-device changing logic
Recent refactors intended for us to pass the new incoming QCameraDevice
as parameter when setting up the new camera session. A typo causes us
to read the current member 'm_cameraDevice' instead of passing the
parameter.

This is not problematic in the current architecture because the values
are the same, but we should use the parameter properly in order for it
to not become problematic for future patches. See relation chain for
context.

Task-number: QTBUG-141376
Change-Id: Ib4549192bd3d5b3d81c3db6e109b494dfb2468a7
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 122706bc36d71e309fa5d098801c598d37e0ba3c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>