Apple, camera: Fix regression in device-switching behavior
Recent refactors to the Apple camera code introduced error-handling
and cleanup during device-switching, in the case where fail to activate
the new device. This logic was implemented with the assumption that
the new device should be configured with the QCameraFormat that is
currently stored. This causes a regression in existing QCamera
behavior whenever we try switching to a device that does not support
the currently select QCameraFormat.

Pre-existing QCamera behavior is that camera-switching should always
imply resetting the QCameraFormat to what is default for the new
device. The old code had less error-handling and so this was made
possible by the QCamera implementation always calling 'setCamera()' and
'setCameraFormat()' in succession.

This patch fixes the regression by explicitly configuring the new
camera-device with the default format, so the backend can do so in a
single operation.

Pick-to: 6.10
Fixes: QTBUG-141376
Change-Id: Ib4736a494968a50eb5b75914416c4b6649004cac
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>