Stop configuring relevant manual tests when building for mobile
The manual tests 'mediadevices' and 'mediaformats' are command line
tools that will error out if you try to configure them when building
for mobile. This stops us from building all examples as a single job,
using ninja.

This patch makes them not be included in the build, when building on
mobile.

Pick-to: 6.10
Change-Id: Iffd0da0d470037551381b45cef5a021babeaf63e
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>