Android, audio: Move JNI class declarations to dedicated header
The Android audio code in the QtMultimedia CMake target uses the Java
class 'QtAudioDeviceManager'. This Java class is declared as a Qt JNI
class in multiple .cpp files. This breaks unity builds because there
will be redefinitions of the same generated Java class.

This patch moves the declaration into a dedicated header.

As a drive-by, we sort some header-includes in the touched .cpp files.

Pick-to: 6.10
Change-Id: I9bb4b38b762963de4d19753a3a66c3948eae465d
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>