Android, FFmpeg: Move QtVideoDeviceManager JNI decl into header
The FFmpeg backend declares the Qt JNI class for 'QtVideoDeviceManager'
in multiple .cpp files. This causes issues during unity builds,
because multiple files will declare the same types.

This patch moves the declaration into a common header. In the future,
this header should be used for other JNI class declarations as well.

Pick-to: 6.10
Change-Id: I5b4dca577e18981b740d8cf1b3c74e4963952413
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>