tst_qvideoframebackend: Test more backends when converting to image
We are able to set the QRhi backend of the thread-local instance used
in QVideoFrame to QImage conversion. This can be used in tests testing
the conversion to increase our coverage.

This patch expands the test running the conversion on a worker thread to
test with all available QRhi backends, as this was a blind spot with
likely failures.

Pick-to: 6.8
Task-number: QTBUG-135598
Change-Id: I6290650fdadf4c8881b3ae51a141fcd804ca1020
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit ec166f8d17d4a0edcd9b8db79115a73fcaab7f1b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>