mediaframeinputs: Fix compile issue
In C++17 it is not legal to explicitly capture 'this' in a lambda when
capturing by copy by default. Meanwhile, in C++20 it is deprecated
to capture 'this' by default. We need this manual test to compile
in both contexts.

This patch fixes it by introducing #ifdefs depending on whether we are
compiling as C++20

This issue blocks us from compiling manual tests in CI. See relation
chain.

Pick-to: 6.10
Change-Id: Ib49b464eee318117803738dea9b7010f7a8949ad
Reviewed-by: Bror Wetlesen Vedeld <bror.vedeld@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>