mediaframeinputs: Fix compile issue
In C++17 it is not legal to explicitly capture 'this' in a lambda when
capturing by copy by default. Meanwhile, in C++20 it is deprecated
to capture 'this' by default. We need this manual test to compile
in both contexts.

This patch fixes it by introducing #ifdefs depending on whether we are
compiling as C++20

This issue blocks us from compiling manual tests in CI. See relation
chain.

Change-Id: Ib49b464eee318117803738dea9b7010f7a8949ad
Reviewed-by: Bror Wetlesen Vedeld <bror.vedeld@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit a364e57d5802ce1864257b3a07fa079e1066ce76)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>