iOS, FindFFmpeg: Look for shared libs in 'lib' subfolder
The FindFFmpeg.cmake script currently looks for iOS FFmpeg shared
libraries in a 'frameworks' subfolder. Incoming patches will make us
expect the shared libraries to exist in the 'lib' subfolder, similar to
how it's done on other platforms.

This patch is a temporary measure that makes us scan the two subfolders
'lib' and 'frameworks' for the relevant shared library files, when
deploying FFmpeg.

Tas-number: QTBUG-136480
Pick-to: 6.10
Change-Id: Ie3f1a4cbf50d32e8b79ca764882c95e7b84d39f6
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>