QQuickImageCapture - register preview image per instance
QQuickImagePreviewProvider only provided a single preview image. This
will break when people use multiple instances in their applications.

We now keep track of the QQuickImageCapture instances and keep on
preview image per instance. This means that we also need to change the
URI to take the instance ID into account.

Change-Id: I85694dff67ff643564fcdd130839b047102c3382
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>