Doc: Capitalize PlaybackOptions QML type
The following QDoc warning 'No output generated for QML property
'PlaybackOptions::probeSize' because 'PlaybackOptions' is undocumented'
assumes that there is a an undocumented PlaybackOptions QML type.

There is, however, a documented playbackOptions \qmltype. This patch
changes the \qmltype name to PlaybackOptions, which is consistent with
QML capitalization conventions. This name change will also make the QDoc warning go away.

Pick-to: 6.10
Change-Id: I60820e3aeb8e0e1bc59b73db21842ca15f3d77fb
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>