GStreamer: Fix crashes and deadlocks during MediaPlayer destruction
In order to switch out our video and audio sinks with fakesinks during
destruction we need to call directly into GstPlay's playbin element.
Doing this while playbin is in the middle of state transition or
pipeline reconfiguration can cause crashes and deadlocks.

By keeping track of expected STOPPED state changed messages from
GstPlay, we can wait for those to appear on its message bus before
doing such calls in ~QGstreamerMediaPlayer. In addition we use a flag
to determine if we need to call gst_play_stop(), or if the pipeline is
already or about to be stopped.

Note: There is one known edge case that can make us wait for a STOPPED
message that never arrives and hit the 200ms timeout in the destructor:
If stop() or setSource() is called after reaching EOS, but before we've
processed the EOS message (e.g. in response to the last positionChanged
signal), m_expectedStoppedMessages is incremented twice, but GstPlay
will typically only post one STOPPED state change message. Accepting
this is safer than risking undercounting.

Fixes: QTBUG-140805
Pick-to: 6.12 6.11 6.8
Change-Id: I579bfa4f3ec0d3097705550cb89d800723ce24f9
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>