Audio: add PulseAudio server detection before starting Pipewire
Arch linux and derivatives allow users to install PipeWire and
PulseAudio in parallel. This is problematic, as there is no good way to
determine if PipeWire manages audio devices from PipeWire's API.

This patch adds a rather ugly workaround to first try to connect to
PulseAudio and detect if it's PipeWire's emulation. If we detect vanilla
PulseAudio, we skip pipewire. Unfortunately this will increase the
startup times by a few milliseconds.

Fixes: QTBUG-141836
Fixes: QTBUG-140938
Pick-to: 6.10
Change-Id: I7cc1257212805c7413dba5a3ba9d9638e170df03
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>