Android: Convert audio sample format to float if format unsupported
We report supporting all sample formats Qt supports on Android, as
we can do sample format conversion. AAudio supports formats except
uint8, but in some cases other formats such as int32 can be unsupported
by Android mixer. There seems to be no way to directly query this
support, only what the audio device itself supports, which is not
equivalent.

This patch adds a fallback mechanism to perform sample format conversion
to float if opening a stream with the original format fails, but opening
it with float succeeds.

Fixes: QTBUG-141852
Pick-to: 6.10
Change-Id: I0b0524ddd2712beeb3a7ecdd37928544610b4a4a
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>