qmlvideo: Stop specifying NSPrincipalClass manually
The example 'qmlvideo', when on Apple platforms, specifies the value
for 'NSPrincipalClass' in the manifest. It is set to always be
'NSApplication' regardless whether we are on macOS or iOS. This causes
a crash on iOS because this type does not exist on that platform.

This patch solves this by removing the value, so that we can lean on
default values. In this case, the default value is already
'NSApplication' for macOS, and it is 'UIApplication' for iOS.

Task-number: QTBUG-141662
Pick-to: 6.10
Change-Id: I1dafdd124d100b44b113a5d870c3a832a58c3ed8
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>