QRtAudioEngine: Give a notification role to our sinks
At the moment we'd get the stream classified as Music (which is the
default) even though QRtAudioEngine seems to be used mostly for
non-music purposes (i.e. sound effects). Declare it as such so the
system isn't entirely verbose over these.
This would mean processes using QtVirtualKeyboard get registered as
Music players, which isn't intended.

Pick-to: 6.10
Change-Id: I34c1f613b1e61c0b2674da16cf12a652e3a4ae8e
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>