QPlatformSurfaceCapture: Add frame rate changing
This patch adds methods to QPlatformSurfaceCapture that allow setting
and getting the target surface capture frame rate, as well as
implementations for using the target rate on backends that utilize a
"pulling" grabber that runs on a timer that we control.

On Windows, despite the grabbing running on a timer, we don't get a new
frame if nothing has changed, making the rate a maximum rate of a
variable rate capture.

Fixes: QTBUG-141822
Fixes: QTBUG-141823
Change-Id: I774300eee68f06c18062adb23160ac550540eab4
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>