QVideoWidget: improve error handling for eglfs
QVideoWidget is not really supported on eglfs, since it creates a
subwindow. This issue came up multiple times by flickering (6.10 and
earlier) as well as with fatal errors (6.11 and newer).
So we update the QVideoWidget to create a centered label instead to
inform users that the QVideoWidget is not supported.

Fixes: QTBUG-116767
Fixes: QTBUG-142542
Pick-to: 6.11 6.10 6.8
Change-Id: I299eacfe780dbff1e92c9e293a396701fb54eaf1
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>