GStreamer: Ensure relay sink doesn't update contexts with a null QRhi
QGstreamerRelayVideoSink updated the GStreamer contexts with the QRhi
instance of the rendering sink from the connected pluggable sink upon
connection. If the rhi was null at this point, previous contexts were
still reset including pointers to the EGL display and target texture.
This caused a sink being added to a media player when its rhi was not
initialized yet to not configure the mediaplayer pipeline correctly.

This patch ensures that the relay video sink updates contexts again if
the pluggable sink reports it's rhi has changed. The thread-local rhi is
used if no rendering rhi exists. The logic for this is streamlined in
this patch as well.

Pick-to: 6.11 6.10 6.8
Change-Id: I801cb1c820652c5418a750f6a183f1f609b9f6b4
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>