WASAPI: fix shutdown of wasapi sink streams
When stop()ing a stream, the stream would keep on running, as we did not
set the stop request when we want to drain the ringbuffer. This caused
the stream to get "idle", but never to terminate.
We should set the stop requested flag for any shutdown policy, though.

Pick-to: 6.10
Change-Id: Idbde00636911a8e29a59a4c6b95e6caee2dd589e
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
(cherry picked from commit ccdc3000307196038e996676e9d043e79c035b25)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>