Android: Set audio device channel config
For Android audio devices we set the channel count limits, but not the
channel configuration. This broke QMediaPlayer downmixing surround audio
to stereo with the FFmpeg backend.

This patch adds a channel configuration to the preferred audio format
for Android QAudioDevices, and uses it as the canonical device channel
configuration.

Fixes: QTBUG-142931
Pick-to: 6.11 6.10
Change-Id: I4072c8eefc7dec8e14601723b880bcb839578f22
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>