QSoundEffect: harden device changes
If we cannot create an engine for the format, we should keep the sample
around, but set the state to QSoundEffect::Error. Simarly, if selecting
a new device fails, we should go to QSoundEffect::Error or else go to
the ready/loading/null state that we would be expected to be in.

Task-number: QTBUG-142743
Pick-to: 6.10
Change-Id: I19b6c59488b73f202cb5217eb2f4dbc669c9a0f9
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 332b731d5b39e971a244e21b3110fb5817d3859b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>