Make a virtual camera addition and removal test cases
We do not have any virtual camera management tests in the Qt Multimedia
tests. This commit adds two test cases, one that adds a virtual camera
using the VCamManager utility, checks that the camera is visible and has
the expected properties, with QMediaDevices, and another that removes
the camera and checks that it no longer exists.
And a struct VCamParameters is added to encapsulate the parameters
needed to create a virtual camera.

Change-Id: I52503cd9368d0e1e8e8d01e793153adda807b9ef
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>