AVFImageCapture: Fix use-after-free error in completion handler
When using Obj-C block functions, local reference variables are
captured as references, not as copies of their referenced type. This
caused an issue in the still photo completion handler where we were
reading a string after it was freed.

This patch fixes this by making a local copy and capturing that
instead.

Pick-to: 6.11 6.10 6.8
Fixes: QTBUG-143058
Change-Id: Id56d20e75e5241925e0825953a98beb033723773
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>